+++ src/domain/challenge/challenge/layout/ChallengePageBanner.tsx
@@ -1,5 +1,5 @@
 import React, { FC } from 'react';
-import EntityPageBanner from '../../../shared/components/PageHeader/EntityPageBanner';
+import JourneyPageBanner from '../../../shared/components/PageHeader/JourneyPageBanner';
 import { useChallenge } from '../hooks/useChallenge';
 import { getVisualBanner } from '../../../common/visual/utils/visuals.utils';
 
@@ -7,12 +7,12 @@ const ChallengePageBanner: FC = () => {
   const { challenge, loading } = useChallenge();
 
   return (
-    <EntityPageBanner
+    <JourneyPageBanner
       title={challenge?.displayName}
       tagline={challenge?.context?.tagline}
       loading={loading}
       bannerUrl={getVisualBanner(challenge?.context?.visuals)}
-      entityTypeName="challenge"
+      journeyTypeName="challenge"
       showBreadcrumbs
     />
   );
