+++ src/domain/platform/constants/EntityTypeName.ts
@@ -1,1 +1,3 @@
-export type EntityTypeName = 'hub' | 'challenge' | 'opportunity' | 'organization' | 'user';
+import { JourneyTypeName } from '../../challenge/JourneyTypeName';
+
+export type EntityTypeName = JourneyTypeName | 'organization' | 'user';
